requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cmake==3.26.4
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@94113be6e12db36b8c7601e13747587f19ec92fe
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fairscale==0.4.12
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.2
fvcore==0.1.5.post20220512
ghp-import==2.1.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.50.0
huggingface-hub==0.16.4
hydra-core==1.2.0
idna==3.4
imageio==2.31.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.9
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.2
jupyter_client==7.4.5
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
KITTI-Iterator==1.4.0
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
memray==1.4.1
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocstrings==0.19.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.57.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.3
portalocker==2.6.0
positional-encodings==6.0.1
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydot==1.4.2
Pygments==2.13.0
pymdown-extensions==9.8
pyparsing==3.0.9
pyrsistent==0.19.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==2.0.4
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.3.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.11.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2023.7.4
timm==0.6.11
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.0
torch-scatter==2.0.9
torchaudio==0.13.0
torchmetrics==1.0.0
torchvision==0.14.0
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.30.2
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.7
yacs==0.1.8
yarl==1.9.2
zipp==3.10.0
